Provide information on parent links existing
This provides information on whether or not we have information on a
parent link existing.  This means that if there's information on a
parent of the current page, we can render some information about
that.  A common case would be when you catalog the top-level page of a
multi-page article.